initial vscodium rust setup
Currently, this will install rustup and gcc as home packages.
It sounds like nix-shell is the preferred and better way to do this
on NixOS, see https://discourse.nixos.org/t/how-do-i-install-rust/7491/8.
I will use this set up for now, and probably revisit when I have a better
understanding of nix-shell and maybe nix-direnv.