feat: home-manager IntelliSense in vscodium
I was using nil, and I thought that it was providing auto-complete for
home-manager through evaluating flake inputs, but that was not true. I
have managed to get this working using someone's config repo on github.
I have not figured out a good way to reference my own
homeConfigurations.<name>.options, because I'm using home-manager as a
NixOS module. I will reference the repo I am using here in the future
to try to create or otherwise access my own homeConfigurations options
set, but until then this is working.